+++ packages/cubejs-server/src/command/dev-server.ts
@@ -21,7 +21,7 @@ export class DevServer extends Command {
     const container = new ServerContainer({
       debug: options.flags.debug,
     });
-    container.runProjectDiagnostics();
+    await container.runProjectDiagnostics();
 
     const configuration = await container.lookupConfiguration();
     container.runServerInstance(configuration);